Rename to SwitchingLanguageDataProvider, add docs
