+++ src/Sitecore.FakeDb/Sitecore.FakeDb.csproj
@@ -77,7 +77,7 @@
     <Compile Include="Configuration\ConfigReader.cs" />
     <Compile Include="Configuration\Settings.cs" />
     <Compile Include="ContentSearch\SwitchingSearchProvider.cs" />
-    <Compile Include="Data\DataProviders\SwitchingLanguagesDataProvider.cs" />
+    <Compile Include="Data\DataProviders\SwitchingLanguageDataProvider.cs" />
     <Compile Include="Data\Engines\DataCommands\Prototypes\AddFromTemplateCommandPrototype.cs" />
     <Compile Include="Data\Engines\DataCommands\Prototypes\AddVersionCommandProtoype.cs" />
     <Compile Include="Data\Engines\DataCommands\Prototypes\BlobStreamExistsCommandPrototype.cs" />
