+++ src/Sitecore.FakeDb/Sitecore.config
@@ -61,7 +61,7 @@
       <main type="Sitecore.FakeDb.Data.DataProviders.$(database)DataProvider, Sitecore.FakeDb">
         <CacheOptions.DisableAll>true</CacheOptions.DisableAll>
       </main>
-      <language type="Sitecore.FakeDb.Data.DataProviders.SwitchingLanguagesDataProvider, Sitecore.FakeDb" />
+      <language type="Sitecore.FakeDb.Data.DataProviders.SwitchingLanguageDataProvider, Sitecore.FakeDb" />
     </dataProviders>
     <!-- DATABASES -->
     <databases>
